auth/requirements.txt
asgiref==3.7.2
blinker==1.6.3
certifi==2023.7.22
click==8.1.7
distlib==0.3.7
Django==4.2.7
filelock==3.13.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
nltk==3.5
packaging==23.2
platformdirs==3.11.0
pluggy==1.3.0
protobuf==4.21.12
PyJWT==2.8.0
pytest==7.4.3
redis==5.0.1
regex==2023.10.3
rq==1.15.1
SQLAlchemy==2.0.22
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
virtualenv==20.24.6
Werkzeug==3.0.1
